adding seed file, updating routing activitySpec
This commit adds a seed file for tests so that we can avoid
duplication in before hooks. There's also a single after hook
which removes all data in the test database. All tests
have been made pending except for the tests to check
the route for GET to /users/:u_id/activities. These tests
have been modified to make use of the seed file. Please note
that data set in the seed file isn't guaranteed to come in
a particular order, so watch out if your expect statements
require a specific order for documents.